+++ packages/deepsea-tools/src/utils/assignFnName.ts
@@ -5,7 +5,7 @@ import { AnyFunction } from "./createFnWithMiddleware"
  * @param target 目标函数
  * @param source 源函数
  */
-export function assignFnName<T extends AnyFunction>(target: T, source: Function | string): T {
+export function assignFnName<T extends AnyFunction>(target: T, source: AnyFunction | string): T {
     const name = typeof source === "string" ? source : source.name
     Object.defineProperty(target, "name", { value: name })
     return target