+++ packages/deepsea-tools/src/utils/compareProperties.ts
@@ -5,7 +5,7 @@ import equal from "fast-deep-equal"
  * @param keyList - 比较的 key 集合
  * @returns 是否相等
  */
-export function compareProperties<T extends Object>(a: T, b: T, ...keyList: (keyof T)[]): boolean {
+export function compareProperties<T extends object>(a: T, b: T, ...keyList: (keyof T)[]): boolean {
     if (keyList.length === 0) throw new Error(`keyList 为空`)
     return keyList.every(key => equal(a[key], b[key]))
 }