+++ packages/soda-array/src/index.ts
@@ -1,6 +1,6 @@
 /** 为数组添加方法 */
 export function extendArrayPrototype() {
-    if (!Array.prototype.hasOwnProperty("with")) {
+    if (!Object.hasOwn(Array.prototype, "with")) {
         class A {
             static with<T>(this: T[], index: number, value: T): T[] {
                 if (!Number.isInteger(index) || index >= this.length || index < this.length * -1) throw new RangeError(`Invalid index : ${index}`)
@@ -13,7 +13,7 @@ export function extendArrayPrototype() {
         Array.prototype.with = A.with
     }
 
-    if (!Array.prototype.hasOwnProperty("toReversed")) {
+    if (!Object.hasOwn(Array.prototype, "toReversed")) {
         function toReversed<T>(this: T[]): T[] {
             const $ = this.slice()
             $.reverse()
@@ -23,7 +23,7 @@ export function extendArrayPrototype() {
         Array.prototype.toReversed = toReversed
     }
 
-    if (!Array.prototype.hasOwnProperty("toShifted")) {
+    if (!Object.hasOwn(Array.prototype, "toShifted")) {
         function toShifted<T>(this: T[]): T[] {
             const $ = this.slice()
             $.shift()
@@ -33,7 +33,7 @@ export function extendArrayPrototype() {
         Array.prototype.toShifted = toShifted
     }
 
-    if (!Array.prototype.hasOwnProperty("toPopped")) {
+    if (!Object.hasOwn(Array.prototype, "toPopped")) {
         function toPopped<T>(this: T[]): T[] {
             const $ = this.slice()
             $.pop()
@@ -43,7 +43,7 @@ export function extendArrayPrototype() {
         Array.prototype.toPopped = toPopped
     }
 
-    if (!Array.prototype.hasOwnProperty("toSorted")) {
+    if (!Object.hasOwn(Array.prototype, "toSorted")) {
         function toSorted<T>(this: T[], compareFn?: (a: T, b: T) => number): T[] {
             const $ = this.slice()
             $.sort(compareFn)
@@ -53,7 +53,7 @@ export function extendArrayPrototype() {
         Array.prototype.toSorted = toSorted
     }
 
-    if (!Array.prototype.hasOwnProperty("toDeduplicated")) {
+    if (!Object.hasOwn(Array.prototype, "toDeduplicated")) {
         function toDeduplicated<T>(this: T[], compareFn?: (a: T, b: T) => boolean): T[] {
             return this.reduce((prev: T[], item: T) => {
                 if (compareFn ? !prev.some(it => compareFn(it, item)) : !prev.includes(item)) prev.push(item)
@@ -64,7 +64,7 @@ export function extendArrayPrototype() {
         Array.prototype.toDeduplicated = toDeduplicated
     }
 
-    if (!Array.prototype.hasOwnProperty("toSpliced")) {
+    if (!Object.hasOwn(Array.prototype, "toSpliced")) {
         function toSpliced<T>(this: T[], start: number, deleteCount?: number, ...items: T[]): T[] {
             const $ = this.slice()
 
@@ -77,7 +77,7 @@ export function extendArrayPrototype() {
         Array.prototype.toSpliced = toSpliced
     }
 
-    if (!Array.prototype.hasOwnProperty("toPushed")) {
+    if (!Object.hasOwn(Array.prototype, "toPushed")) {
         function toPushed<T>(this: T[], ...items: T[]): T[] {
             const $ = this.slice()
             $.push(...items)
@@ -87,7 +87,7 @@ export function extendArrayPrototype() {
         Array.prototype.toPushed = toPushed
     }
 
-    if (!Array.prototype.hasOwnProperty("toUnshifted")) {
+    if (!Object.hasOwn(Array.prototype, "toUnshifted")) {
         function toUnshifted<T>(this: T[], ...items: T[]): T[] {
             const $ = this.slice()
             $.unshift(...items)
@@ -97,7 +97,7 @@ export function extendArrayPrototype() {
         Array.prototype.toUnshifted = toUnshifted
     }
 
-    if (!Array.prototype.hasOwnProperty("toExchange")) {
+    if (!Object.hasOwn(Array.prototype, "toExchange")) {
         function toExchange<T>(this: T[], a: number, b: number): T[] {
             return this.with(a, this[b]).with(b, this[a])
         }
@@ -105,7 +105,7 @@ export function extendArrayPrototype() {
         Array.prototype.toExchange = toExchange
     }
 
-    if (!Array.prototype.hasOwnProperty("at")) {
+    if (!Object.hasOwn(Array.prototype, "at")) {
         function at<T>(this: T[], index: number): T | undefined {
             if (!Number.isInteger(index)) throw new RangeError(`Invalid index : ${index}`)
             return this[index >= 0 ? index : this.length + index]
@@ -114,7 +114,7 @@ export function extendArrayPrototype() {
         Array.prototype.at = at
     }
 
-    if (!Array.prototype.hasOwnProperty("chunk")) {
+    if (!Object.hasOwn(Array.prototype, "chunk")) {
         Array.prototype.chunk = function chunk<T>(this: T[], size: number) {
             if (!Number.isInteger(size) || size <= 0) throw new RangeError(`Invalid size : ${size}`)
             return this.reduce((prev: T[][], item: T, index: number) => {
@@ -126,31 +126,31 @@ export function extendArrayPrototype() {
         }
     }
 
-    if (!Array.prototype.hasOwnProperty("nonNullable")) {
+    if (!Object.hasOwn(Array.prototype, "nonNullable")) {
         Array.prototype.nonNullable = function nonNullable<T>(this: T[]) {
             return this.filter(item => item !== undefined && item !== null)
         }
     }
 
-    if (!Array.prototype.hasOwnProperty("asNonNullable")) {
+    if (!Object.hasOwn(Array.prototype, "asNonNullable")) {
         Array.prototype.asNonNullable = function asNonNullable<T>(this: T[]) {
             return this
         }
     }
 
-    if (!Array.prototype.hasOwnProperty("difference")) {
+    if (!Object.hasOwn(Array.prototype, "difference")) {
         Array.prototype.difference = function difference<T>(this: T[], values: T[], compareFn?: (a: T, b: T) => boolean) {
             return this.filter(item => (compareFn ? !values.some(it => compareFn(it, item)) : !values.includes(item)))
         }
     }
 
-    if (!Array.prototype.hasOwnProperty("intersection")) {
+    if (!Object.hasOwn(Array.prototype, "intersection")) {
         Array.prototype.intersection = function intersection<T>(this: T[], values: T[], compareFn?: (a: T, b: T) => boolean) {
             return this.filter(item => (compareFn ? values.some(it => compareFn(it, item)) : values.includes(item)))
         }
     }
 
-    if (!Array.prototype.hasOwnProperty("union")) {
+    if (!Object.hasOwn(Array.prototype, "union")) {
         Array.prototype.union = function union<T>(this: T[], values: T[], compareFn?: (a: T, b: T) => boolean) {
             return this.concat(
                 values.reduce((prev: T[], item) => {
@@ -161,7 +161,7 @@ export function extendArrayPrototype() {
         }
     }
 
-    if (!Array.prototype.hasOwnProperty("random")) {
+    if (!Object.hasOwn(Array.prototype, "random")) {
         Array.prototype.random = function random<T>(this: Array<T>, start?: number, end?: number) {
             const arr = this.slice(start, end)
             if (arr.length === 0) throw new Error("cant random an empty array")