+++ packages/deepsea-components/CHANGELOG.md
@@ -1,5 +1,14 @@
 # deepsea-components
 
+## 5.17.13
+
+### Patch Changes
+
+- 优化代码
+- Updated dependencies
+    - deepsea-tools@5.43.4
+    - soda-hooks@6.16.3
+
 ## 5.17.12
 
 ### Patch Changes