+++ packages/react-soda/CHANGELOG.md
@@ -1,5 +1,11 @@
 # react-soda
 
+## 5.3.5
+
+### Patch Changes
+
+- 优化代码
+
 ## 5.3.4
 
 ### Patch Changes