+++ packages/soda-coordinate/CHANGELOG.md
@@ -1,5 +1,11 @@
 # soda-coordinate
 
+## 5.3.5
+
+### Patch Changes
+
+- 优化代码
+
 ## 5.3.4
 
 ### Patch Changes