+++ packages/soda-react-router/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-react-router
 
+## 1.2.8
+
+### Patch Changes
+
+- Updated dependencies
+    - soda-hooks@6.16.3
+
 ## 1.2.7
 
 ### Patch Changes