+++ packages/soda-tanstack-form/CHANGELOG.md
@@ -1,5 +1,11 @@
 # soda-tanstack-form
 
+## 0.3.5
+
+### Patch Changes
+
+- 优化代码
+
 ## 0.3.4
 
 ### Patch Changes